+++ crates/wast/src/ast.rs
@@ -1,0 +1,1 @@
+
