+++ crates/wast/src/core/global.rs
@@ -1,20 +1,22 @@
-use crate::ast::{self, kw};
+use crate::core::*;
+use crate::kw;
 use crate::parser::{Parse, Parser, Result};
+use crate::token::{Id, NameAnnotation, Span};
 
 /// A WebAssembly global in a module
 #[derive(Debug)]
 pub struct Global<'a> {
     /// Where this `global` was defined.
-    pub span: ast::Span,
+    pub span: Span,
     /// An optional name to reference this global by
-    pub id: Option<ast::Id<'a>>,
+    pub id: Option<Id<'a>>,
     /// An optional name for this function stored in the custom `name` section.
-    pub name: Option<ast::NameAnnotation<'a>>,
+    pub name: Option<NameAnnotation<'a>>,
     /// If present, inline export annotations which indicate names this
     /// definition should be exported under.
-    pub exports: ast::InlineExport<'a>,
+    pub exports: InlineExport<'a>,
     /// The type of this global, both its value type and whether it's mutable.
-    pub ty: ast::GlobalType<'a>,
+    pub ty: GlobalType<'a>,
     /// What kind of global this defined as.
     pub kind: GlobalKind<'a>,
 }
@@ -27,10 +29,10 @@ pub enum GlobalKind<'a> {
     /// ```text
     /// (global i32 (import "foo" "bar"))
     /// ```
-    Import(ast::InlineImport<'a>),
+    Import(InlineImport<'a>),
 
     /// A global defined inline in the module itself
-    Inline(ast::Expression<'a>),
+    Inline(Expression<'a>),
 }
 
 impl<'a> Parse<'a> for Global<'a> {
