+++ crates/wast/src/core/module.rs
@@ -1,50 +1,19 @@
-use crate::ast::{self, annotation, kw};
+use crate::core::*;
 use crate::parser::{Parse, Parser, Result};
+use crate::token::{Id, IndexOrRef, ItemRef, NameAnnotation, Span};
+use crate::{annotation, kw};
 
-pub use crate::resolve::Names;
-
-/// A `*.wat` file parser, or a parser for one parenthesized module.
-///
-/// This is the top-level type which you'll frequently parse when working with
-/// this crate. A `*.wat` file is either one `module` s-expression or a sequence
-/// of s-expressions that are module fields.
-#[derive(Debug)]
-pub struct Wat<'a> {
-    #[allow(missing_docs)]
-    pub module: Module<'a>,
-}
-
-impl<'a> Parse<'a> for Wat<'a> {
-    fn parse(parser: Parser<'a>) -> Result<Self> {
-        if !parser.has_meaningful_tokens() {
-            return Err(parser.error("expected at least one module field"));
-        }
-        let _r = parser.register_annotation("custom");
-        let module = if !parser.peek2::<kw::module>() {
-            let fields = ModuleField::parse_remaining(parser)?;
-            Module {
-                span: ast::Span { offset: 0 },
-                id: None,
-                name: None,
-                kind: ModuleKind::Text(fields),
-            }
-        } else {
-            parser.parens(|parser| parser.parse())?
-        };
-        module.validate(parser)?;
-        Ok(Wat { module })
-    }
-}
+pub use crate::core::resolve::Names;
 
 /// A parsed WebAssembly module.
 #[derive(Debug)]
 pub struct Module<'a> {
     /// Where this `module` was defined
-    pub span: ast::Span,
+    pub span: Span,
     /// An optional identifier this module is known by
-    pub id: Option<ast::Id<'a>>,
+    pub id: Option<Id<'a>>,
     /// An optional `@name` annotation for this module
-    pub name: Option<ast::NameAnnotation<'a>>,
+    pub name: Option<NameAnnotation<'a>>,
     /// What kind of module this was parsed as.
     pub kind: ModuleKind<'a>,
 }
@@ -71,9 +40,9 @@ impl<'a> Module<'a> {
     /// where expansion and name resolution happens.
     ///
     /// This function will mutate the AST of this [`Module`] and replace all
-    /// [`super::Index`] arguments with `Index::Num`. This will also expand inline
-    /// exports/imports listed on fields and handle various other shorthands of
-    /// the text format.
+    /// [`Index`](crate::token::Index) arguments with `Index::Num`. This will
+    /// also expand inline exports/imports listed on fields and handle various
+    /// other shorthands of the text format.
     ///
     /// If successful the AST was modified to be ready for binary encoding. A
     /// [`Names`] structure is also returned so if you'd like to do your own
@@ -84,7 +53,7 @@ impl<'a> Module<'a> {
     /// If an error happens during resolution, such a name resolution error or
     /// items are found in the wrong order, then an error is returned.
     pub fn resolve(&mut self) -> std::result::Result<Names<'a>, crate::Error> {
-        crate::resolve::resolve(self)
+        crate::core::resolve::resolve(self)
     }
 
     /// Encodes this [`Module`] to its binary form.
@@ -113,10 +82,10 @@ impl<'a> Module<'a> {
     /// expansion-related errors.
     pub fn encode(&mut self) -> std::result::Result<Vec<u8>, crate::Error> {
         self.resolve()?;
-        Ok(crate::binary::encode(self))
+        Ok(crate::core::binary::encode(self))
     }
 
-    fn validate(&self, parser: Parser<'_>) -> Result<()> {
+    pub(crate) fn validate(&self, parser: Parser<'_>) -> Result<()> {
         let mut starts = 0;
         if let ModuleKind::Text(fields) = &self.kind {
             for item in fields.iter() {
@@ -162,22 +131,22 @@ impl<'a> Parse<'a> for Module<'a> {
 #[allow(missing_docs)]
 #[derive(Debug)]
 pub enum ModuleField<'a> {
-    Type(ast::Type<'a>),
-    Import(ast::Import<'a>),
-    Func(ast::Func<'a>),
-    Table(ast::Table<'a>),
-    Memory(ast::Memory<'a>),
-    Global(ast::Global<'a>),
-    Export(ast::Export<'a>),
-    Start(ast::ItemRef<'a, kw::func>),
-    Elem(ast::Elem<'a>),
-    Data(ast::Data<'a>),
-    Tag(ast::Tag<'a>),
-    Custom(ast::Custom<'a>),
+    Type(Type<'a>),
+    Import(Import<'a>),
+    Func(Func<'a>),
+    Table(Table<'a>),
+    Memory(Memory<'a>),
+    Global(Global<'a>),
+    Export(Export<'a>),
+    Start(ItemRef<'a, kw::func>),
+    Elem(Elem<'a>),
+    Data(Data<'a>),
+    Tag(Tag<'a>),
+    Custom(Custom<'a>),
 }
 
 impl<'a> ModuleField<'a> {
-    fn parse_remaining(parser: Parser<'a>) -> Result<Vec<ModuleField>> {
+    pub(crate) fn parse_remaining(parser: Parser<'a>) -> Result<Vec<ModuleField>> {
         let mut fields = Vec::new();
         while !parser.is_empty() {
             fields.push(parser.parens(ModuleField::parse)?);
@@ -211,7 +180,7 @@ impl<'a> Parse<'a> for ModuleField<'a> {
         }
         if parser.peek::<kw::start>() {
             parser.parse::<kw::start>()?;
-            return Ok(ModuleField::Start(parser.parse::<ast::IndexOrRef<_>>()?.0));
+            return Ok(ModuleField::Start(parser.parse::<IndexOrRef<_>>()?.0));
         }
         if parser.peek::<kw::elem>() {
             return Ok(ModuleField::Elem(parser.parse()?));
