+++ crates/wast/src/wast.rs
@@ -1,6 +1,8 @@
-use crate::ast::{self, kw};
+use crate::core::{Expression, Module};
+use crate::kw;
 use crate::parser::{Cursor, Parse, Parser, Peek, Result};
-use crate::{AssertExpression, NanPattern, V128Pattern};
+use crate::token::{Id, Span};
+use crate::{AssertExpression, NanPattern, V128Pattern, Wat};
 
 /// A parsed representation of a `*.wast` file.
 ///
@@ -24,7 +26,7 @@ impl<'a> Parse<'a> for Wast<'a> {
                 directives.push(parser.parens(|p| p.parse())?);
             }
         } else {
-            let module = parser.parse::<ast::Wat>()?.module;
+            let module = parser.parse::<Wat>()?.module;
             directives.push(WastDirective::Module(module));
         }
         Ok(Wast { directives })
@@ -54,56 +56,56 @@ impl Peek for WastDirectiveToken {
 #[allow(missing_docs)]
 #[derive(Debug)]
 pub enum WastDirective<'a> {
-    Module(ast::Module<'a>),
+    Module(Module<'a>),
     QuoteModule {
-        span: ast::Span,
+        span: Span,
         source: Vec<&'a [u8]>,
     },
     AssertMalformed {
-        span: ast::Span,
+        span: Span,
         module: QuoteModule<'a>,
         message: &'a str,
     },
     AssertInvalid {
-        span: ast::Span,
+        span: Span,
         module: QuoteModule<'a>,
         message: &'a str,
     },
     Register {
-        span: ast::Span,
+        span: Span,
         name: &'a str,
-        module: Option<ast::Id<'a>>,
+        module: Option<Id<'a>>,
     },
     Invoke(WastInvoke<'a>),
     AssertTrap {
-        span: ast::Span,
+        span: Span,
         exec: WastExecute<'a>,
         message: &'a str,
     },
     AssertReturn {
-        span: ast::Span,
+        span: Span,
         exec: WastExecute<'a>,
-        results: Vec<ast::AssertExpression<'a>>,
+        results: Vec<AssertExpression<'a>>,
     },
     AssertExhaustion {
-        span: ast::Span,
+        span: Span,
         call: WastInvoke<'a>,
         message: &'a str,
     },
     AssertUnlinkable {
-        span: ast::Span,
-        module: ast::Module<'a>,
+        span: Span,
+        module: Module<'a>,
         message: &'a str,
     },
     AssertException {
-        span: ast::Span,
+        span: Span,
         exec: WastExecute<'a>,
     },
 }
 
 impl WastDirective<'_> {
     /// Returns the location in the source that this directive was defined at
-    pub fn span(&self) -> ast::Span {
+    pub fn span(&self) -> Span {
         match self {
             WastDirective::Module(m) => m.span,
             WastDirective::AssertMalformed { span, .. }
@@ -270,9 +272,9 @@ impl<'a> Parse<'a> for WastDirective<'a> {
 #[derive(Debug)]
 pub enum WastExecute<'a> {
     Invoke(WastInvoke<'a>),
-    Module(ast::Module<'a>),
+    Module(Module<'a>),
     Get {
-        module: Option<ast::Id<'a>>,
+        module: Option<Id<'a>>,
         global: &'a str,
     },
 }
@@ -299,10 +301,10 @@ impl<'a> Parse<'a> for WastExecute<'a> {
 #[allow(missing_docs)]
 #[derive(Debug)]
 pub struct WastInvoke<'a> {
-    pub span: ast::Span,
-    pub module: Option<ast::Id<'a>>,
+    pub span: Span,
+    pub module: Option<Id<'a>>,
     pub name: &'a str,
-    pub args: Vec<ast::Expression<'a>>,
+    pub args: Vec<Expression<'a>>,
 }
 
 impl<'a> Parse<'a> for WastInvoke<'a> {
@@ -326,7 +328,7 @@ impl<'a> Parse<'a> for WastInvoke<'a> {
 #[allow(missing_docs)]
 #[derive(Debug)]
 pub enum QuoteModule<'a> {
-    Module(ast::Module<'a>),
+    Module(Module<'a>),
     Quote(Vec<&'a [u8]>),
 }
 
@@ -348,8 +350,8 @@ impl<'a> Parse<'a> for QuoteModule<'a> {
 
 #[cfg(test)]
 mod tests {
-    use crate::ast::wast::WastDirective;
     use crate::parser::{parse, ParseBuffer};
+    use crate::wWastDirective;
 
     macro_rules! assert_parses_to_directive {
         ($text:expr, $pattern:pat) => {{
